Gets top two trends, searches, prints
Now the code gets the top two trends from google trends related to a search term, searches them in Youtube, and prints the details of the top videos for those terms to user. Now going to add a check for views vs. subs, since that will be a great indicator of a good topic